+++ config/html/critical.js
@@ -1,0 +1,10 @@
+/**
+ * critical configuration.
+ *
+ * @since unreleased
+ *
+ * @type {Object}
+ */
+module.exports = {
+  inline: true
+}