+++ src/pshry.com/data/masthead.js
@@ -1,5 +1,12 @@
 const site = require('./site')
 
+/**
+ * Masthead data object.
+ *
+ * @since unreleased
+ *
+ * @type {Object}
+ */
 module.exports = {
   title: site.title,
 }